+++ src/structure_factor/utils.py
@@ -1,10 +1,8 @@
 """Collection of secondary functions used in the principal modules."""
 
 import numpy as np
-from scipy import interpolate, stats
+from scipy import stats
 from scipy.special import j0, j1, jn_zeros, jv, y0, y1, yv
-
-# utils for point_processes.py
 
 
 def get_random_number_generator(seed=None):
@@ -12,7 +10,52 @@ def get_random_number_generator(seed=None):
     return np.random.default_rng(seed)
 
 
-# utils for hyperuniformity.py
+def set_nan_inf_to_zero(array, nan=0, posinf=0, neginf=0):
+    """Set nan, posinf, and neginf values of ``array`` to the corresponding input arguments. Defaults to zero."""
+    return np.nan_to_num(array, nan=nan, posinf=posinf, neginf=neginf)
+
+
+def bessel1(order, x):
+    """Evaluate `Bessel function of the first kind <https://en.wikipedia.org/wiki/Bessel_function>`_."""
+    if order == 0:
+        return j0(x)
+    if order == 1:
+        return j1(x)
+    return jv(order, x)
+
+
+def bessel1_zeros(order, nb_zeros):
+    """Evaluate zeros of the `Bessel function of the first kind <https://en.wikipedia.org/wiki/Bessel_function>`_."""
+    return jn_zeros(order, nb_zeros)
+
+
+def bessel2(order, x):
+    """Evaluate `Bessel function of the second kind <https://en.wikipedia.org/wiki/Bessel_function>`_."""
+    if order == 0:
+        return y0(x)
+    if order == 1:
+        return y1(x)
+    return yv(order, x)
+
+
+def norm(k):
+    return np.linalg.norm(k, axis=-1)
+
+
+def meshgrid_to_column_matrix(X):
+    r"""Transform output ``X`` of numpy.meshgrid to a 2d numpy array with columns formed by flattened versions of the elements of ``X``.
+
+    .. code-block:: python
+
+        np.column_stack([x.ravel() for x in X])
+
+    Args:
+        X (list): output of numpy.meshgrid.
+
+    Returns:
+        np.ndarray: 2d array.
+    """
+    return np.column_stack([x.ravel() for x in X])
 
 
 def sort_by_keys(keys, *arrays, **argsort_params):
@@ -44,9 +87,6 @@ def _sort_vectors(k, x_k, y_k=None):
     return sort_by_keys(k, k, x_k, y_k)
 
 
-# utils for hyperuniformity.py and structure_factor.py
-
-
 def _bin_statistics(x, y, **params):
     """Divide ``x`` into bins and evaluate the mean and the standard deviation of the corresponding elements of ``y`` over each bin.
 
@@ -71,88 +111,3 @@ def _bin_statistics(x, y, **params):
     bin_std, _, _ = stats.binned_statistic(x, y, statistic="std", **params)
     bin_std /= np.sqrt(count)
     return bin_centers, bin_mean, bin_std
-
-
-# utils for tranform.py
-
-
-def bessel1(order, x):
-    """Evaluate `Bessel function of the first kind <https://en.wikipedia.org/wiki/Bessel_function>`_."""
-    if order == 0:
-        return j0(x)
-    if order == 1:
-        return j1(x)
-    return jv(order, x)
-
-
-def bessel1_zeros(order, nb_zeros):
-    """Evaluate zeros of the `Bessel function of the first kind <https://en.wikipedia.org/wiki/Bessel_function>`_."""
-    return jn_zeros(order, nb_zeros)
-
-
-def bessel2(order, x):
-    """Evaluate `Bessel function of the second kind <https://en.wikipedia.org/wiki/Bessel_function>`_."""
-    if order == 0:
-        return y0(x)
-    if order == 1:
-        return y1(x)
-    return yv(order, x)
-
-
-# utils for pair_correlation_function.py
-
-
-def set_nan_inf_to_zero(array, nan=0, posinf=0, neginf=0):
-    """Set nan, posinf, and neginf values of ``array`` to the corresponding input arguments. Defaults to zero."""
-    return np.nan_to_num(array, nan=nan, posinf=posinf, neginf=neginf)
-
-
-def _extrapolate_pcf(x, r, pcf_r, **params):
-    """Interpolate pcf_r for x=<r_max and set to 1 for x>r_max.
-
-    Args:
-        x (numpy.ndarray): Points on which the pair correlation function is to be evaluated.
-
-        r (numpy.ndarray): Vector of the radius on with the pair correlation function was evaluated.
-
-        pcf_r (numpy.ndarray): Vector of evaluations of the pair correlation function corresponding to ``r``.
-
-    Returns:
-        numpy.ndarray: evaluation of the extrapolated pair correlation function on ``x``.
-    """
-    r_max = np.max(r)  # maximum radius
-    pcf = np.zeros_like(x)
-    params.setdefault("fill_value", "extrapolate")
-
-    mask = x > r_max
-    if np.any(mask):
-        pcf[mask] = 1.0
-        np.logical_not(mask, out=mask)
-        pcf[mask] = interpolate.interp1d(r, pcf_r, **params)(x[mask])
-    else:
-        pcf = interpolate.interp1d(r, pcf_r, **params)(x)
-
-    return pcf
-
-
-# utils for structure_factor.py
-
-
-def norm(k):
-    return np.linalg.norm(k, axis=-1)
-
-
-def meshgrid_to_column_matrix(X):
-    r"""Transform output ``X`` of numpy.meshgrid to a 2d numpy array with columns formed by flattened versions of the elements of ``X``.
-
-    .. code-block:: python
-
-        np.column_stack([x.ravel() for x in X])
-
-    Args:
-        X (list): output of numpy.meshgrid.
-
-    Returns:
-        np.ndarray: 2d array.
-    """
-    return np.column_stack([x.ravel() for x in X])
